Criando a classe Cartao

+++ src/main/java/mycash/modelo/Banco.java
@@ -4,17 +4,17 @@ import java.util.Collection;
 
 public class Banco {
 
-	private int id;
+	private Long id;
 
 	private int nome;
 
 	private Collection<Conta> contas;
-
-	public int getId() {
+	
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
